requirements.txt
vosk
pyaudio
python-dotenv
baidu-aip
